Unify AuthenticationRequestParameters
