+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/AuthenticationRequestParameters.kt
@@ -2,6 +2,9 @@ package at.asitplus.openid
 
 import at.asitplus.KmmResult.Companion.wrap
 import at.asitplus.dif.PresentationDefinition
+import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
+import at.asitplus.signum.indispensable.io.ByteArrayBase64UrlSerializer
+import at.asitplus.signum.indispensable.josef.JsonWebToken
 import at.asitplus.signum.indispensable.josef.io.InstantLongSerializer
 import kotlinx.datetime.Instant
 import kotlinx.serialization.SerialName
@@ -258,10 +261,186 @@ data class AuthenticationRequestParameters(
      */
     @SerialName("code_challenge_method")
     val codeChallengeMethod: String? = null,
+
+    /**
+     * CSC: Optional
+     * Request a preferred language according to RFC 5646
+     */
+    @SerialName("lang")
+    val lang: String? = null,
+
+    /**
+     * CSC: REQUIRED-"credential"
+     * The identifier associated to the credential to authorize.
+     * This parameter value may contain characters that are reserved, unsafe or
+     * forbidden in URLs and therefore SHALL be url-encoded by the signature
+     * application
+     */
+    @SerialName("credentialID")
+    @Serializable(ByteArrayBase64UrlSerializer::class)
+    val credentialID: ByteArray? = null,
+
+    /**
+     * CSC: Required-"credential"
+     * This parameter contains the symbolic identifier determining the kind of
+     * signature to be created
+     */
+    @SerialName("signatureQualifier")
+    val signatureQualifier: SignatureQualifier? = null,
+
+    /**
+     * CSC: Required-"credential"
+     * The number of signatures to authorize
+     */
+    @SerialName("numSignatures")
+    val numSignatures: Int? = null,
+
+    /**
+     * CSC: REQUIRED-"credential"
+     * One or more base64url-encoded hash values to be signed
+     */
+    @SerialName("hashes")
+    @Serializable(HashesSerializer::class)
+    val hashes: Hashes? = null,
+
+    /**
+     * CSC: REQUIRED-"credential"
+     * String containing the OID of the hash algorithm used to generate the hashes
+     */
+    @SerialName("hashAlgorithmOID")
+    val hashAlgorithmOid: ObjectIdentifier? = null,
+
+    /**
+     * CSC: OPTIONAL
+     * A free form description of the authorization transaction in the lang language.
+     * The maximum size of the string is 500 characters
+     */
+    @SerialName("description")
+    val description: String? = null,
+
+    /**
+     * CSC: OPTIONAL
+     * To restrict access to the authorization server of a remote service, this specification introduces the
+     * additional account_token parameter to be used when calling the oauth2/authorize endpoint. This
+     * parameter contains a secure token designed to authenticate the authorization request based on an
+     * Account ID that SHALL be uniquely assigned by the signature application to the signing user or to the
+     * user’s application account
+     */
+    @SerialName("account_token")
+    val accountToken: JsonWebToken? = null,
+
+    /**
+     * CSC: OPTIONAL
+     * Arbitrary data from the signature application. It can be used to handle a
+     * transaction identifier or other application-spe cific data that may be useful for
+     * debugging purposes
+     */
+    @SerialName("clientData")
+    val clientData: String? = null,
+
+    /**
+     * OID4VP: OPTIONAL. Array of strings, where each string is a base64url encoded JSON object that contains a typed
+     * parameter set with details about the transaction that the Verifier is requesting the End-User to authorize.
+     * The Wallet MUST return an error if a request contains even one unrecognized transaction data type or transaction
+     * data not conforming to the respective type definition.
+     */
+    @SerialName("transaction_data")
+    override val transactionData: Set<String>? = null,
 ) : RequestParameters {
 
     fun serialize() = odcJsonSerializer.encodeToString(this)
-    override val transactionData: Set<String>? = null
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null || this::class != other::class) return false
+
+        other as AuthenticationRequestParameters
+
+        if (numSignatures != other.numSignatures) return false
+        if (responseType != other.responseType) return false
+        if (clientId != other.clientId) return false
+        if (redirectUrl != other.redirectUrl) return false
+        if (scope != other.scope) return false
+        if (state != other.state) return false
+        if (nonce != other.nonce) return false
+        if (claims != other.claims) return false
+        if (clientMetadata != other.clientMetadata) return false
+        if (clientMetadataUri != other.clientMetadataUri) return false
+        if (idTokenHint != other.idTokenHint) return false
+        if (request != other.request) return false
+        if (requestUri != other.requestUri) return false
+        if (idTokenType != other.idTokenType) return false
+        if (presentationDefinition != other.presentationDefinition) return false
+        if (presentationDefinitionUrl != other.presentationDefinitionUrl) return false
+        if (authorizationDetails != other.authorizationDetails) return false
+        if (clientIdScheme != other.clientIdScheme) return false
+        if (walletIssuer != other.walletIssuer) return false
+        if (userHint != other.userHint) return false
+        if (issuerState != other.issuerState) return false
+        if (responseMode != other.responseMode) return false
+        if (responseUrl != other.responseUrl) return false
+        if (audience != other.audience) return false
+        if (issuer != other.issuer) return false
+        if (issuedAt != other.issuedAt) return false
+        if (resource != other.resource) return false
+        if (codeChallenge != other.codeChallenge) return false
+        if (codeChallengeMethod != other.codeChallengeMethod) return false
+        if (lang != other.lang) return false
+        if (credentialID != null) {
+            if (other.credentialID == null) return false
+            if (!credentialID.contentEquals(other.credentialID)) return false
+        } else if (other.credentialID != null) return false
+        if (signatureQualifier != other.signatureQualifier) return false
+        if (!hashes.contentEquals(other.hashes)) return false
+        if (hashAlgorithmOid != other.hashAlgorithmOid) return false
+        if (description != other.description) return false
+        if (accountToken != other.accountToken) return false
+        if (clientData != other.clientData) return false
+        if (transactionData != other.transactionData) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        var result = numSignatures ?: 0
+        result = 31 * result + (responseType?.hashCode() ?: 0)
+        result = 31 * result + (clientId?.hashCode() ?: 0)
+        result = 31 * result + (redirectUrl?.hashCode() ?: 0)
+        result = 31 * result + (scope?.hashCode() ?: 0)
+        result = 31 * result + (state?.hashCode() ?: 0)
+        result = 31 * result + (nonce?.hashCode() ?: 0)
+        result = 31 * result + (claims?.hashCode() ?: 0)
+        result = 31 * result + (clientMetadata?.hashCode() ?: 0)
+        result = 31 * result + (clientMetadataUri?.hashCode() ?: 0)
+        result = 31 * result + (idTokenHint?.hashCode() ?: 0)
+        result = 31 * result + (request?.hashCode() ?: 0)
+        result = 31 * result + (requestUri?.hashCode() ?: 0)
+        result = 31 * result + (idTokenType?.hashCode() ?: 0)
+        result = 31 * result + (presentationDefinition?.hashCode() ?: 0)
+        result = 31 * result + (presentationDefinitionUrl?.hashCode() ?: 0)
+        result = 31 * result + (authorizationDetails?.hashCode() ?: 0)
+        result = 31 * result + (clientIdScheme?.hashCode() ?: 0)
+        result = 31 * result + (walletIssuer?.hashCode() ?: 0)
+        result = 31 * result + (userHint?.hashCode() ?: 0)
+        result = 31 * result + (issuerState?.hashCode() ?: 0)
+        result = 31 * result + (responseMode?.hashCode() ?: 0)
+        result = 31 * result + (responseUrl?.hashCode() ?: 0)
+        result = 31 * result + (audience?.hashCode() ?: 0)
+        result = 31 * result + (issuer?.hashCode() ?: 0)
+        result = 31 * result + (issuedAt?.hashCode() ?: 0)
+        result = 31 * result + (resource?.hashCode() ?: 0)
+        result = 31 * result + (codeChallenge?.hashCode() ?: 0)
+        result = 31 * result + (codeChallengeMethod?.hashCode() ?: 0)
+        result = 31 * result + (lang?.hashCode() ?: 0)
+        result = 31 * result + (credentialID?.contentHashCode() ?: 0)
+        result = 31 * result + (signatureQualifier?.hashCode() ?: 0)
+        result = 31 * result + (hashes?.contentHashCode() ?: 0)
+        result = 31 * result + (hashAlgorithmOid?.hashCode() ?: 0)
+        result = 31 * result + (description?.hashCode() ?: 0)
+        result = 31 * result + (accountToken?.hashCode() ?: 0)
+        result = 31 * result + (clientData?.hashCode() ?: 0)
+        result = 31 * result + (transactionData?.hashCode() ?: 0)
+        return result
+    }
 
     companion object {
         fun deserialize(it: String) = kotlin.runCatching {
