+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/AuthorizationDetails.kt
@@ -1,7 +1,7 @@
 import at.asitplus.openid.AuthorizationDetails
 import at.asitplus.rqes.collection_entries.DocumentLocation
 import at.asitplus.rqes.collection_entries.OAuthDocumentDigest
-import at.asitplus.rqes.enums.SignatureQualifier
+import at.asitplus.openid.SignatureQualifier
 import at.asitplus.signum.indispensable.asn1.ObjectIdSerializer
 import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
 import kotlinx.serialization.SerialName
