+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/serializers/RequestParameterSerializer.kt
@@ -1,7 +1,6 @@
 package at.asitplus.rqes.serializers
 
 import at.asitplus.openid.AuthenticationRequestParameters
-import at.asitplus.openid.CscAuthenticationRequestParameters
 import at.asitplus.openid.RequestParameters
 import at.asitplus.rqes.SignatureRequestParameters
 import kotlinx.serialization.DeserializationStrategy
@@ -13,8 +12,7 @@ object RequestParametersSerializer : JsonContentPolymorphicSerializer<RequestPar
     override fun selectDeserializer(element: JsonElement): DeserializationStrategy<RequestParameters> {
         val parameters = element.jsonObject
         return when {
-            "signatureQualifier" in parameters -> SignatureRequestParameters.serializer()
-            "hashes" in parameters -> CscAuthenticationRequestParameters.serializer()
+            "documentDigests" in parameters -> SignatureRequestParameters.serializer()
             else -> AuthenticationRequestParameters.serializer()
         }
     }
